+++ src/Service/Lambda/src/Result/ListVersionsByFunctionResponse.php
@@ -22,6 +22,7 @@ use AsyncAws\Lambda\ValueObject\LoggingConfig;
 use AsyncAws\Lambda\ValueObject\RuntimeVersionConfig;
 use AsyncAws\Lambda\ValueObject\RuntimeVersionError;
 use AsyncAws\Lambda\ValueObject\SnapStartResponse;
+use AsyncAws\Lambda\ValueObject\TenancyConfig;
 use AsyncAws\Lambda\ValueObject\TracingConfigResponse;
 use AsyncAws\Lambda\ValueObject\VpcConfigResponse;
 
@@ -232,6 +233,7 @@ class ListVersionsByFunctionResponse extends Result implements \IteratorAggregat
             'SnapStart' => empty($json['SnapStart']) ? null : $this->populateResultSnapStartResponse($json['SnapStart']),
             'RuntimeVersionConfig' => empty($json['RuntimeVersionConfig']) ? null : $this->populateResultRuntimeVersionConfig($json['RuntimeVersionConfig']),
             'LoggingConfig' => empty($json['LoggingConfig']) ? null : $this->populateResultLoggingConfig($json['LoggingConfig']),
+            'TenancyConfig' => empty($json['TenancyConfig']) ? null : $this->populateResultTenancyConfig($json['TenancyConfig']),
         ]);
     }
 
@@ -376,6 +378,13 @@ class ListVersionsByFunctionResponse extends Result implements \IteratorAggregat
         }
 
         return $items;
+    }
+
+    private function populateResultTenancyConfig(array $json): TenancyConfig
+    {
+        return new TenancyConfig([
+            'TenantIsolationMode' => (string) $json['TenantIsolationMode'],
+        ]);
     }
 
     private function populateResultTracingConfigResponse(array $json): TracingConfigResponse